nn/pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "noise-detect"
version = "0.1.0"
description = "Electric pump audio classifier (PyTorch) for future ESP32-S3 deployment."
readme = "README.md"
requires-python = ">=3.13"
keywords = ["audio", "classification", "embedded", "esp32", "pytorch", "ml", "signal-processing"]
authors = [{ name = "g1ibby" }]
license = { text = "MIT" }
dependencies = [
  "torch>=2.3,<3",
  "torchaudio>=2.3,<3",
  "torchmetrics>=1.4,<2",
  "lightning>=2.3,<3",
  "hydra-core>=1.3,<2",
  "omegaconf>=2.3,<3",
  "numpy>=1.26",
  "soundfile>=0.12",
  "torch-audiomentations>=0.11",
  "rich>=13.7",
  "tqdm>=4.66",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dashboard = [
  # Aim depends on aimrocks which currently lacks cp313 wheels.
  # Restrict dashboard extra to Python < 3.13 to avoid resolver failures.
  "aim>=3.21; python_version < '3.13'",
]
dev = [
  "pytest>=8.2",
  "mypy>=1.10",
  "ruff>=0.5",
  "black>=24.8",
  "pyright>=1.1",
  "types-PyYAML>=6.0.12",
  "ipykernel>=6.29",
  "ipywidgets>=8.1",
]
export = [
  # ONNX toolchain for host-side export (quantization runs in Docker)
  "onnx>=1.19",
  "onnxruntime>=1.23.1",
  "onnxslim>=0.1.70",
]

[tool.hatch.build.targets.wheel]
packages = ["src/noise_detect"]

[tool.black]
line-length = 100
target-version = ["py313"]
include = "(src|tests|scripts)/"
exclude = '''
/(\.git|\.hg|\.mypy_cache|\.ruff_cache|\.venv|build|dist|runs)/
'''

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["src", "tests", "scripts"]
exclude = [
  ".git",
  ".hg",
  ".mypy_cache",
  ".ruff_cache",
  ".venv",
  "build",
  "dist",
  "runs",
]

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "W",   # pycodestyle warnings
  "I",   # isort
  "UP",  # pyupgrade
  "N",   # pep8-naming
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "PL",  # pylint (subset)
  "PT",  # flake8-pytest-style
  "RUF", # ruff-specific rules
  "SIM", # flake8-simplify
  "ANN", # flake8-annotations (typing)
  "TCH", # flake8-type-checking (TYPE_CHECKING)
]
ignore = [
  "ANN101",  # missing type annotation for self
  "ANN102",  # missing type annotation for cls
]
fix = true

[tool.ruff.lint.isort]
known-first-party = ["noise_detect"]
combine-as-imports = true

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]
"tests/**" = [
  "ANN",      # relax annotations in tests
  "PLR2004",  # magic value used in comparison
]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
show_error_codes = true
pretty = true
files = ["src/noise_detect", "scripts", "tests"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-ra -q"
filterwarnings = [
  "ignore::UserWarning",
]
